requirements-py2.txt
atomicwrites==1.3.0
attrs==18.2.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
googleapis-common-protos==1.5.8
grpcio==1.18.0
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.8.1
protobuf==3.6.1
py==1.7.0
pytest==4.3.0
scandir==1.9.0
six==1.12.0
